refactor: lift layout up to browser/ssr level
