+++ gatsby-ssr.js
@@ -5,3 +5,11 @@
  */
 
 // You can delete this file if you're not using it
+
+const React = require('react')
+const Layout = require('./src/components/Layout')
+
+exports.wrapRootElement = props => {
+    const { element } = props
+    return <Layout>{element}</Layout>
+}
