+++ src/pages/index.jsx
@@ -1,18 +1,18 @@
 import React from 'react'
-import { Layout, SEO } from '../components'
+import { SEO } from '../components'
 import LandingBio from '../components/LandingBio'
 import { useSiteMetadata } from '../hooks'
 
 function MainIndex() {
-    const {title, description} = useSiteMetadata()
+    const { title, description } = useSiteMetadata()
     return (
-        <Layout>
+        <>
             <SEO
                 title={title}
                 keywords={[`blog`, `gatsby`, `javascript`, `react`]}
             />
-            <LandingBio title={title} description={description}/>
-        </Layout>
+            <LandingBio title={title} description={description} />
+        </>
     )
 }
