+++ src/pages/lists.jsx
@@ -1,14 +1,13 @@
 import React from 'react'
 import { graphql } from 'gatsby'
-import { Bio,ColumnLinkWrapper, Layout, SEO, ListedLink } from '../components'
-import useSiteMetadata from '../hooks/useSiteMetadata'
+import { Bio,ColumnLinkWrapper,  SEO, ListedLink } from '../components'
 
 function Lists(props) {
     const { data } = props
     const list = data.lists.edges
 
     return (
-        <Layout>
+        <>
             <SEO title="Lists" keywords={['reading', 'books']} />
             <h1>List</h1>
             <p>A list of lists I maintain.</p>
@@ -27,7 +26,7 @@ function Lists(props) {
                 </ColumnLinkWrapper>
 
             <Bio />
-        </Layout>
+        </>
     )
 }
