+++ src/templates/BlogList.js
@@ -1,43 +1,37 @@
 import React from 'react'
 import { graphql } from 'gatsby'
-import styled from 'styled-components'
+
 import {
     Bio,
     BlogExcerpt,
-    Layout,
     PageNavigation,
     SEO,
     Search,
     StyledLink,
 } from '../components'
 
-// TODO: figure out if this is used/helpful - think it conflicts with the <layout>
-const Content = styled.div`
-    margin: 0 auto;
-    max-width: 860px;
-    padding: 1.45rem 1.0875rem;
-`
-
 function BlogList(props) {
     const { data } = props
     const { previousPage, nextPage } = props.pageContext
     const posts = data.allMarkdownRemark.edges
     return (
-        <Layout>
+        <>
             <SEO
                 title="All posts"
                 keywords={[`blog`, `gatsby`, `javascript`, `react`]}
             />
-            <Content>
-                <h1><StyledLink to={"/blog"}>/*Code-Comments*/</StyledLink></h1>
-                <Search />
-                {posts.map(({ node }) => (
-                    <BlogExcerpt key={node.fields.slug} node={node} />
-                ))}
-                <PageNavigation previous={previousPage} next={nextPage} />
-            </Content>
+            <h1>
+                <StyledLink to={'/blog'}>/*Code-Comments*/</StyledLink>
+            </h1>
+            <PageNavigation previous={previousPage} next={nextPage} />
+            <Search />
+            {posts.map(({ node }) => (
+                <BlogExcerpt key={node.fields.slug} node={node} />
+            ))}
+            <PageNavigation previous={previousPage} next={nextPage} />
+
             <Bio />
-        </Layout>
+        </>
     )
 }
